pyproject.toml
[tool.poetry]
name = "frazerport"
version = "0.1.0"
description = "this is a anti boredom allience bot"
authors = ["Daylight inc."]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.0"
pyuca = "^1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
